pyproject.toml
[tool.poetry]
name = "django-imagedeck"
version = "0.3.2"
description = "A package to seamlessly group different sorts of images into groups."
authors = ["Robert Turnbull <robert.turnbull@unimelb.edu.au>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://gitlab.unimelb.edu.au/rturnbull/django-imagedeck"
packages = [
    { include = "imagedeck" },
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
django-imagekit = "^4.0.2"
django-polymorphic = "^3.0.0"
easy-thumbnails = "^2.7.1"
django-filer = "^2.0.2"
requests = "^2.26.0"
iiif-prezi = "^0.3.0"

[tool.poetry.dev-dependencies]
coverage = "^5.5"
black = "^21.9b0"

[build-system]
requires = ["setuptools >= 40.6.0", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
